Backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.1.2
requests==2.32.0
python-dotenv==1.1.0
